requirements.txt
Pillow
requests
matplotlib
cryptography